Move guideline classes to General
